requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django
gunicorn
django-etc==1.4.0
django-hitcount==1.3.5
django-jazzmin==2.6.0
django-modeltranslation==0.18.7
django_debug_toolbar==3.8.1
Pillow==9.4.0
pycodestyle==2.10.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
whitenoise
